Add transcription to sms
Was able to get the transcription working for sms.
The message now has the number from the caller, the category they
picked to call about, the recording url, and the transcription message.
I added an action to the recording, so if it will loop to the beginning
of the message if no one hangsup.